\name{runSeqsumm}
\alias{runSeqsumm}
\title{
Function that generates seqsumm summarized files.
}
\description{
The function calls the seqsumm program to summarize aligned reads into genomic windows. The results are output to a file called \code{seqsumm_out.txt}.
}
\usage{
runSeqsumm(summ.win=50, file=NULL, folder=NULL, output.file="seqsumm_out.txt", samtools.path="samtools")
}

\arguments{
  \item{summ.win}{
The genomic window size, expressed in kilobases as a whole positive number, in which to summarize the reads.
}
  \item{file}{
Full path to the SAM or BAM file.
It is recommended to place it alone in a folder, where analysis output files will be generated.
}
  \item{folder}{
Only used if folder is not indicated.
Path to the folder where the sample SAM or BAM is located.
The function first searches for the SAM file and, if not existent, for the BAM file.
If several SAM (or BAM) files exist in the folder, only the first one will be used.
The summarization results will be written to this same folder.
}
  \item{output.file}{
The name of the file where the output will be stored (in the same folder as the SAM/BAM file).
Be aware, if you choose an alterative name, that function \code{\link{readSeqsumm}} will look for \code{seqsumm_out.txt}.
}
  \item{samtools.path}{
The program SAMtools is required if your alignment file is in BAM format.
In case SAMtools is not in your path variable, you should indicate the path to the executable.
}
}
\details{
Beware that this is a computationally intensive task, so it might take from a few minutes to several hours depending on the number of reads in the aligned file.
}
\value{
A file called \code{seqsumm_out.txt} (or the indicated name), which will be read by the function \code{\link{readSeqsumm}}, is generated in the indicated folder. Therefore, you only need to call this function once, unless you want to generate a new file with other input/parameters.
The output file has chromosome, window, mean GC and mean mapping quality columns, and either a count column for single-end alignments or five count columns for paired-end alignments:
- Type 1. Mapped with correct orientation and within insert size.
- Type 2. Mapped with correct orientation and uniquely, but wrong insert size.
- Type 3. Mapped within the insert size but wrong orientation.
- Type 4. Mapped uniquely, but wrong orientation and insert size.
- Type 5. One of the mates in unmapped.
See \code{\link{seqsumm_HCC1143}} for an example table generated by the function.
The function does not return anything.
}
\author{
David Mosen-Ansorena
}
\examples{
file.copy(file.path(system.file(package="seqCNA"), "extdata/test.sam"), tempdir())
tempdir() 

window.size = 50

runSeqsumm(window.size, folder=tempdir())
}
\keyword{ Summarization }
