\name{seqCNA-package}
\alias{seqCNA-package}
\alias{seqCNA}
\docType{package}
\title{
Copy number analysis of deep sequencing cancer data
}
\description{
Fast summarization, extensive filtering and improved normalization
}
\details{
\tabular{ll}{
Package: \tab seqCNA\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2013-08-03\cr
License: \tab GPL-3\cr
}
applyFilters applyThresholds plotCNProfile readSeqsumm runGLAD runSeqnorm runSeqsumm writeCNProfile
}
\author{
David Mosen-Ansorena

Maintainer: David Mosen-Ansorena <dmosen.gn@cicbiogune.es>
}
